requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.6.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.14
fastjsonschema==2.20.0
filelock==3.16.1
fsspec==2024.9.0
h11>=0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
ipython==8.12.3
jedi==0.19.1
Jinja2>=3.1.6
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core>=5.8.0
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
openai==1.86.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pwinput==1.0.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart>=0.0.18
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.2
requests>=2.32.4
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette>=0.40.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.20.0
tornado>=6.5.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3>=2.5.0
uvicorn==0.30.6
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
yarg==0.1.9
chromadb==1.0.15
markdown2==2.5.3
fpdf==1.7.2
pandas==2.2.2
pgvector
weasyprint